move remote_link_accessors
