+++ corehq/apps/app_manager/management/commands/link_app_to_remote.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import
 from django.core.management import BaseCommand
 
 from corehq.apps.app_manager.models import LinkedApplication, RemoteAppDetails
-from corehq.apps.app_manager.remote_link_accessors import whilelist_app_on_remote
+from corehq.apps.linked_domain.remote_accessors import whilelist_app_on_remote
 
 
 class Command(BaseCommand):
